Minor UI files reorganization
