+++ ui/src/stock/StockController.ts
