+++ ui/src/stock/index.ts
@@ -1,0 +1,3 @@
+export * from "./StockController";
+export * from "./StockRowController";
+export * from "./StockTableController";
